+++ LC-API/Comp/LC_APIManager.cs
@@ -1,6 +1,5 @@
 using LC_API.GameInterfaceAPI;
 using LC_API.GameInterfaceAPI.Features;
-using LC_API.Networking;
 using LC_API.ServerAPI;
 using UnityEngine;