+++ LC-API/GameInterfaceAPI/Events/EventArgs/Player/GrabbedItemEventArgs.cs
@@ -1,14 +1,12 @@
 ﻿namespace LC_API.GameInterfaceAPI.Events.EventArgs.Player
 {
-    public class GrabbingItemEventArgs : System.EventArgs
+    public class GrabbedItemEventArgs : System.EventArgs
     {
         public Features.Player Player { get; }
 
         public Features.Item Item { get; }
 
-        public bool IsAllowed { get; set; } = true;
-
-        public GrabbingItemEventArgs(Features.Player player, Features.Item item)
+        public GrabbedItemEventArgs(Features.Player player, Features.Item item)
         {
             Player = player;
             Item = item;