+++ LC-API/GameInterfaceAPI/Events/Patches/Internal/DisplayTipPatch.cs
@@ -1,8 +1,5 @@
 ﻿using HarmonyLib;
 using LC_API.GameInterfaceAPI.Features;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace LC_API.GameInterfaceAPI.Events.Patches.Internal
 {