+++ LC-API/GameInterfaceAPI/Events/Patches/Player/Hurt.cs
@@ -1,7 +1,6 @@
 ﻿using GameNetcodeStuff;
 using HarmonyLib;
 using LC_API.GameInterfaceAPI.Events.EventArgs.Player;
-using System;
 using System.Collections.Generic;
 using System.Reflection.Emit;
 using UnityEngine;