+++ LC-API/ManualPatches/ServerPatch.cs
@@ -1,11 +1,7 @@
 ﻿using LC_API.Comp;
-using LC_API.Data;
-using LC_API.Extensions;
 using LC_API.ServerAPI;
 using Steamworks;
 using Steamworks.Data;
-using System;
-using System.Linq;
 using UnityEngine;
 using UnityEngine.InputSystem;